Proyecto Terminado 1.0

+++ js/carrito.js
@@ -16,10 +16,9 @@ function cargarProductosCarrito() {
         carritoVacio.classList.add("oculto");
         carritoProductos.classList.remove("oculto");
         carritoAcciones.classList.remove("oculto");
-        carritoCompraFinal.classList.add("oculto");
-    
+
         carritoProductos.innerHTML = "";
-    
+
         productoEnCarrito.forEach(producto => {
             const div = document.createElement("div");
             div.classList.add("carritoProducto");
@@ -48,7 +47,6 @@ function cargarProductosCarrito() {
         carritoVacio.classList.remove("oculto");
         carritoProductos.classList.add("oculto");
         carritoAcciones.classList.add("oculto");
-        carritoCompraFinal.classList.add("oculto");
     };
     actualizarBotonesEliminar();
     actualizarTotal();
@@ -64,6 +62,26 @@ function actualizarBotonesEliminar() {
 };
 
 function eliminarDelCarrito(e) {
+
+    Toastify({
+        text: "Eliminado",
+        duration: 3000,
+        close: true,
+        gravity: "top", 
+        position: "right", 
+        stopOnFocus: true,
+        style: {
+            background: "linear-gradient(to right, #678d58, black)",
+            borderRadius: "2rem",
+            textTransform: "upperCase",
+        },
+        offset: {
+            x: '1.5rem', 
+            y: '1.5rem' 
+        },
+        onClick: function () {}
+    }).showToast();
+
     const idBoton = e.currentTarget.id;
     const index = productoEnCarrito.findIndex(producto => producto.id === idBoton);
     productoEnCarrito.splice(index, 1);
@@ -73,25 +91,68 @@ function eliminarDelCarrito(e) {
 
 botonVaciar.addEventListener("click", vaciarCarrito);
 function vaciarCarrito() {
-    productoEnCarrito.length = 0;
-    localStorage.setItem("productos-en-carrito", JSON.stringify(productoEnCarrito));
-    cargarProductosCarrito();
+
+    Swal.fire({
+        title: "¿Estás seguro?",
+        icon: "question",
+        html: `Se va a vaciar todo el carrito`,
+        showCancelButton: true,
+        focusConfirm: false,
+        confirmButtonText: `Si`,
+        cancelButtonText: `No`,
+    }).then((result) => {
+        if (result.isConfirmed) {
+            productoEnCarrito.length = 0;
+            localStorage.setItem("productos-en-carrito", JSON.stringify(productoEnCarrito));
+            cargarProductosCarrito();
+        };
+    });
 };
+
 
 function actualizarTotal() {
     const totalCalculado = productoEnCarrito.reduce((acc, producto) => acc + (producto.precio * producto.cantidad), 0);
     total.innerText = `$${totalCalculado}`;
 };
 
-botonComprar.addEventListener("click", comprarCarrito);
-function comprarCarrito() {
-    productoEnCarrito.length = 0;
-    localStorage.setItem("productos-en-carrito", JSON.stringify(productoEnCarrito));
-        carritoVacio.classList.add("oculto");
-        carritoProductos.classList.add("oculto");
-        carritoAcciones.classList.add("oculto");
-        carritoCompraFinal.classList.remove("oculto");
-};
+
+botonComprar.addEventListener("click", function () {
+    mostrarFormulario();
+});
+function mostrarFormulario() {
+    Swal.fire({
+        title: 'Completa la compra',
+        html:
+            '<input id="nombre" class="swal2-input" placeholder="Nombre del comprador">' +
+            '<input id="tarjeta" class="swal2-input" placeholder="Número de tarjeta" type="number">',
+        focusConfirm: false,
+        preConfirm: () => {
+            const nombre = Swal.getPopup().querySelector('#nombre').value;
+            const tarjeta = Swal.getPopup().querySelector('#tarjeta').value;
+
+            if (!nombre || !tarjeta) {
+                Swal.showValidationMessage('Completa ambos campos');
+            }
+            
+            const datosCompra = { nombre, tarjeta };
+            console.log(datosCompra);
+
+            productoEnCarrito.length = 0;
+            localStorage.setItem("productos-en-carrito", JSON.stringify(productoEnCarrito))
+            cargarProductosCarrito();
+
+            return datosCompra;
+        }
+    }).then((result) => {
+        if (result.isConfirmed) {
+            Swal.fire('¡Gracias por tu compra!', '', 'success');
+        }
+    });
+}
+
+
+
+/* MODO COLOR */
 
 const colorModeButton = document.querySelector("#colorMode");
 const main = document.querySelector("main");
@@ -115,7 +176,7 @@ if (darkMode === "activado") {
 
 colorModeButton.addEventListener("click", () => {
     darkMode = localStorage.getItem("darkMode");
-    if(darkMode === "activado") {
+    if (darkMode === "activado") {
         desactivarDarkMode();
         colorModeButton.innerText = "Cambiar a DarkMode";
         localStorage.setItem("colorModeButtonText", "Cambiar a DarkMode");
